+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Add `Enumerables.ToDictionary` extension methods.
+- Add `Enumerables.CreateSafetyCopy` extension method.
 
 ## [2.1.0] - 2018-03-23
 - Add `CodePoint.IsSurrogatePair` property.